Refactor Rust Admin
