+++ rust_admin/src/errors.rs
@@ -1,0 +1,43 @@
+use axum::{
+    http::{header, HeaderValue, StatusCode},
+    response::{IntoResponse, Response},
+    Json,
+};
+use serde_json::json;
+
+#[derive(Debug)]
+pub(crate) struct ApiError {
+    pub(crate) status: StatusCode,
+    pub(crate) detail: String,
+    authenticate: bool,
+}
+
+impl ApiError {
+    pub(crate) fn new(status: StatusCode, detail: impl Into<String>) -> Self {
+        Self {
+            status,
+            detail: detail.into(),
+            authenticate: false,
+        }
+    }
+
+    pub(crate) fn unauthorized(detail: impl Into<String>) -> Self {
+        Self {
+            status: StatusCode::UNAUTHORIZED,
+            detail: detail.into(),
+            authenticate: true,
+        }
+    }
+}
+
+impl IntoResponse for ApiError {
+    fn into_response(self) -> Response {
+        let mut response = (self.status, Json(json!({ "detail": self.detail }))).into_response();
+        if self.authenticate {
+            response
+                .headers_mut()
+                .insert(header::WWW_AUTHENTICATE, HeaderValue::from_static("Bearer"));
+        }
+        response
+    }
+}
